added some piping, moved enums and vaiables to own file, added spacing in header fullName variable
