+++ adviseme-webapp/src/app/_shared/models/user.ts
@@ -1,35 +1,5 @@
 import { Injectable } from '@angular/core';
-
-export enum Role {
-  None,
-  Student, // 0
-  Advisor, // 1
-  Admin,   // 2
-}
-
-export enum Status {
-  None,
-  freshman, // 0
-  sophomore, // 1
-  junior,   // 2
-  senior,   // 3
-  sSenior,  // 4 Super Senior
-}
-
-export enum Major {
-  None,
-  AerospaceEng,     // 0
-  BiomedicalEng,    // 1
-  ChemicalEng,      // 2
-  CivilEng,         // 3
-  EnvironmentalEng, // 4
-  ComputerScience,  // 5
-  ComputerEng,      // 6
-  ElectricalEng,    // 7
-  InfoTech,         // 8
-  MechanicalEng,    // 9
-  NuclearEng,       // 10
-}
+import { Role, Status, Major } from './constants';
 // authentication user profile
 @Injectable()
 export class User {
