+++ adviseme-webapp/src/app/pages/profile/profile-view/profile-view.component.html
@@ -9,17 +9,17 @@
 
           <div class="input-group">
             <span class="input-group-addon"><i class="ion-person" aria-hidden="true"></i></span>
-            <span class="form-control">{{currentUser.firstName}}</span>
+            <span class="form-control">{{currentUser.firstName | ngxCapitalize}}</span>
           </div>
 
           <div class="input-group">
             <span class="input-group-addon"><i class="ion-person" aria-hidden="true"></i></span>
-            <span class="form-control">{{currentUser.lastName}}</span>
+            <span class="form-control">{{currentUser.lastName | ngxCapitalize}}</span>
           </div>
 
           <div class="input-group">
             <span class="input-group-addon"><i class="ion-email" aria-hidden="true"></i></span>
-            <span class="form-control">{{emailBegin}}</span>
+            <span class="form-control">{{emailBegin | ngxCapitalize}}</span>
             <span class="input-group-btn">
               <button class="btn btn-primary">
                 {{emailEnding}}
@@ -34,11 +34,11 @@
 
           <div class="input-group">
             <span class="input-group-addon"><i class="ion-university" aria-hidden="true"></i></span>
-            <span class="form-control">{{currentUser.status}}</span>
+            <span class="form-control">{{currentUser.status | ngxCapitalize}}</span>
           </div>
           <div class="input-group">
             <span class="input-group-addon"><i class="ion-university" aria-hidden="true"></i></span>
-            <span class="form-control">{{currentUser.major}}</span>
+            <span class="form-control">{{currentUser.major | ngxCapitalize}}</span>
           </div>
       </form>
       </nb-card-body>
